requirements.txt
asgiref==3.7.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.1
crispy-bootstrap4==2024.10
distlib==0.3.8
Django==4.2.6
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-registration-redux==2.13
djangorestframework==3.15.2
easygui==0.98.3
filelock==3.17.0
idna==3.4
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.10
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-decouple==3.8
python-dotenv==1.0.0
reportlab==4.2.5
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.29.1
whitenoise==6.6.0
